proxmox_domain_rewrite_sync_for_adguard/requirements.txt
python-dotenv
requests
icmplib